Add reboot, start serial and stop serial buttons to test webpage
Also redefine XSHUT_PIN as zero, it was previously set to pin 2 which could have done anything as init turns lidar sensor off then on.

Note: lidar sensor does not work after reboot and this is due to the poweroff/on achieved with XSHUT_PIN does not work as no pin is connecfed to the XSHUT pin on the sensor.